Revert "se modificaro todas las ramas del proyecto"

+++ src/vehiculosinterfases/Motor.java
@@ -3,7 +3,7 @@ package vehiculosinterfases;
 
 
 public interface Motor {
-    double calcularRevolucionesMotor(int fuerza, int radio);
+    int calcularRevolucionesMotor(int fuerza, int radio);
     
     
 }
